June 5 updates
