+++ .gitignore
@@ -183,6 +183,3 @@ geo_env
 
 ./chromedriver_temp
 ./chromedriver
-
-notebooks/searchterms_process/old
-notebooks/old
